zhengd edit

+++ www/js/OrderP-app.js
@@ -1,4 +1,7 @@
 function logout() {
+    if (confirm("是否退出该账户")) {
+        return;
+    }
     window.localStorage.clear();
     location.href = "../html/Login-app.html";
 }
